fix(seo): ensure layout metadata applies across all pages (#185)
* fix(seo): ensure layout metadata applies across all pages by converting client pages to server wrappers\n\n- Convert home, privacy policy, and terms pages to server components (wrappers)\n- Add per-page metadata for privacy policy and terms\n- Move interactive logic into client components (privacy-client, terms-client)\n- Remove redundant page-level metadata from home to inherit from layout\n\nThis fixes missing meta tags on non-blog routes by enabling Next.js App Router metadata inheritance from .

* refactor(privacy): remove redundant isWiderPage property from PrivacyClient component

* refactor(ui): extract scroll fade logic into reusable hook and update PrivacyClient and TermsClient to use it

* fix: add safe back navigation with fallback to home page

- Update back button logic in privacy-policy and terms pages
- Check window.history.length before calling router.back()
- Fallback to router.push('/') for direct arrivals
- Add client-safe window object check

---------
