+++ app/page.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { Suspense } from "react";
 import { Preloader } from "./components";
 import { MainPageContent } from "./components/MainPageContent";
